add terms in Bootstrap

+++ src/main/java/com/mufic/Final/api/v2/model/CourseDTO.java
@@ -28,7 +28,7 @@ public class CourseDTO {
 
     private int force;
 
-    int program;
+    Long program;
 
     String preCourse;
